+++ library/swift/FinalStreamIntel.swift
@@ -2,36 +2,39 @@
 import Foundation
 
 /// Exposes one time HTTP stream metrics, context, and other details.
+/// Note: -1 means "not present" for the fields of type Int64.
 @objcMembers
 public final class FinalStreamIntel: StreamIntel {
   /// The time the request started, in ms since the epoch.
-  public let requestStartMs: UInt64
+  public let requestStartMs: Int64
   /// The time the DNS resolution for this request started, in ms since the epoch.
-  public let dnsStartMs: UInt64
+  public let dnsStartMs: Int64
   /// The time the DNS resolution for this request completed, in ms since the epoch.
-  public let dnsEndMs: UInt64
+  public let dnsEndMs: Int64
   /// The time the upstream connection started, in ms since the epoch. (1)
-  public let connectStartMs: UInt64
+  public let connectStartMs: Int64
   /// The time the upstream connection completed, in ms since the epoch. (1)
-  public let connectEndMs: UInt64
+  public let connectEndMs: Int64
   /// The time the SSL handshake started, in ms since the epoch. (1)
-  public let sslStartMs: UInt64
+  public let sslStartMs: Int64
   /// The time the SSL handshake completed, in ms since the epoch. (1)
-  public let sslEndMs: UInt64
+  public let sslEndMs: Int64
   /// The time the first byte of the request was sent upstream, in ms since the epoch.
-  public let sendingStartMs: UInt64
+  public let sendingStartMs: Int64
   /// The time the last byte of the request was sent upstream, in ms since the epoch.
-  public let sendingEndMs: UInt64
+  public let sendingEndMs: Int64
   /// The time the first byte of the response was received, in ms since the epoch.
-  public let responseStartMs: UInt64
+  public let responseStartMs: Int64
   /// The time the last byte of the request was received, in ms since the epoch.
-  public let requestEndMs: UInt64
+  public let requestEndMs: Int64
   /// True if the upstream socket had been used previously.
   public let socketReused: Bool
   /// The number of bytes sent upstream.
   public let sentByteCount: UInt64
   /// The number of bytes received from upstream.
   public let receivedByteCount: UInt64
+  /// The response flags for the upstream stream.
+  public let responseFlags: UInt64
 
   // NOTE(1): These fields may not be set if socket_reused is false.
 
@@ -39,20 +42,21 @@ public final class FinalStreamIntel: StreamIntel {
     streamId: Int64,
     connectionId: Int64,
     attemptCount: UInt64,
-    requestStartMs: UInt64,
-    dnsStartMs: UInt64,
-    dnsEndMs: UInt64,
-    connectStartMs: UInt64,
-    connectEndMs: UInt64,
-    sslStartMs: UInt64,
-    sslEndMs: UInt64,
-    sendingStartMs: UInt64,
-    sendingEndMs: UInt64,
-    responseStartMs: UInt64,
-    requestEndMs: UInt64,
+    requestStartMs: Int64,
+    dnsStartMs: Int64,
+    dnsEndMs: Int64,
+    connectStartMs: Int64,
+    connectEndMs: Int64,
+    sslStartMs: Int64,
+    sslEndMs: Int64,
+    sendingStartMs: Int64,
+    sendingEndMs: Int64,
+    responseStartMs: Int64,
+    requestEndMs: Int64,
     socketReused: Bool,
     sentByteCount: UInt64,
-    receivedByteCount: UInt64
+    receivedByteCount: UInt64,
+    responseFlags: UInt64
   ) {
     self.requestStartMs = requestStartMs
     self.dnsStartMs = dnsStartMs
@@ -68,6 +72,7 @@ public final class FinalStreamIntel: StreamIntel {
     self.socketReused = socketReused
     self.sentByteCount = sentByteCount
     self.receivedByteCount = receivedByteCount
+    self.responseFlags = responseFlags
     super.init(streamId: streamId, connectionId: connectionId, attemptCount: attemptCount)
   }
 }
@@ -91,7 +96,8 @@ extension FinalStreamIntel {
       requestEndMs: cFinalIntel.request_end_ms,
       socketReused: cFinalIntel.socket_reused != 0,
       sentByteCount: cFinalIntel.sent_byte_count,
-      receivedByteCount: cFinalIntel.received_byte_count
+      receivedByteCount: cFinalIntel.received_byte_count,
+      responseFlags: cFinalIntel.response_flags
     )
   }
 }